requirements.txt
absl-py==1.4.0
array-record==0.5.0
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
etils==1.5.2
exceptiongroup==1.1.3
executing==2.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.4.0
keras==2.14.0
Keras-Applications==1.0.8
keras-vggface @ git+https://github.com/yaledhlab/vggface.git@b76539b7588bca69b0030ad7e4f985f877dc7c0a
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mtcnn==0.1.1
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
